Nueva validacion para Citizen y Position
Las validaciones que se agrego para ambos mantenimientos consiste en que el administrador no podra crear, modificar, o eliminar ninguna posicion o ciudadano si existe una eleccion abierta.

+++ Sistema-de-votacion/Controllers/PositionsController.cs
@@ -9,6 +9,7 @@ using Microsoft.EntityFrameworkCore;
 using Sistema_de_votacion.Data;
 using Sistema_de_votacion.Models;
 using Sistema_de_votacion.Services.Candidates.Positions;
+using Sistema_de_votacion.Services.Elections;
 
 namespace Sistema_de_votacion.Controllers
 {
@@ -16,10 +17,12 @@ namespace Sistema_de_votacion.Controllers
     public class PositionsController : Controller
     {
         private readonly IPositionService _positionService;
+        private readonly IElectionService _electionService;
 
-        public PositionsController(IPositionService positionService)
+        public PositionsController(IPositionService positionService, IElectionService electionService)
         {
             _positionService = positionService;
+            _electionService = electionService;
         }
 
         // GET: Positions
@@ -31,6 +34,11 @@ namespace Sistema_de_votacion.Controllers
         // GET: Positions/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (await _electionService.VerifyElectionOpen())
+            {
+                ViewBag.Message="No es posible Modificar ninguna posicion politica porque actualmente existe una eleccion abierta.";
+                return RedirectToAction("Index");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -47,8 +55,13 @@ namespace Sistema_de_votacion.Controllers
         }
 
         // GET: Positions/Create
-        public IActionResult Create()
+        public async Task<IActionResult> Create()
         {
+            if (await _electionService.VerifyElectionOpen())
+            {
+                ViewBag.Message = "No es posible crear o añadir ninguna posicion politica porque actualmente existe una eleccion abierta.";
+                return RedirectToAction("Index");
+            }
             return View("Form");
         }
 
@@ -117,6 +130,11 @@ namespace Sistema_de_votacion.Controllers
         // GET: Positions/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (await _electionService.VerifyElectionOpen())
+            {
+                ViewBag.Message = "No es posible Eliminar ninguna posicion politica porque actualmente existe una eleccion abierta.";
+                return RedirectToAction("Index");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -137,6 +155,11 @@ namespace Sistema_de_votacion.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (await _electionService.VerifyElectionOpen())
+            {
+                ViewBag.Message = "No es posible eliminar ninguna posicion politica porque actualmente existe una eleccion abierta.";
+                return RedirectToAction("Index");
+            }
             var position = await _positionService.GetPositionById(id);
             await  _positionService.DeletePosition(position);
